BAEL-17391 Slice 13 | The top 100 articles should have their own package in the module
- Moved folders to respective packages
