+++ core-java-modules/core-java-8/src/test/java/com/baeldung/java_8_features/groupingby/Java8GroupingByCollectorUnitTest.java
@@ -1,4 +1,4 @@
-package com.baeldung.java8;
+package com.baeldung.java_8_features.groupingby;
 
 import com.baeldung.java_8_features.groupingby.BlogPost;
 import com.baeldung.java_8_features.groupingby.BlogPostType;
