+++ jackson-simple/src/test/java/com/baeldung/jackson/ignore/dtos/MyDtoIgnoreFieldByName.java
@@ -1,25 +1,16 @@
-package com.baeldung.jackson.dtos.ignore;
+package com.baeldung.jackson.ignore.dtos;
 
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.annotation.JsonInclude.Include;
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
-@JsonInclude(Include.NON_NULL)
-public class MyDtoIgnoreNull {
+@JsonIgnoreProperties(value = { "intValue" })
+public class MyDtoIgnoreFieldByName {
 
     private String stringValue;
     private int intValue;
     private boolean booleanValue;
 
-    public MyDtoIgnoreNull() {
+    public MyDtoIgnoreFieldByName() {
         super();
-    }
-
-    public MyDtoIgnoreNull(final String stringValue, final int intValue, final boolean booleanValue) {
-        super();
-
-        this.stringValue = stringValue;
-        this.intValue = intValue;
-        this.booleanValue = booleanValue;
     }
 
     // API
